claudecode/universal-scraper/requirements.txt
requests>=2.28.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
lxml>=4.9.0
playwright>=1.40.0
python-docx>=0.8.11
markdownify>=0.11.0
